long/long long -> int64_t, etc

+++ dolen_common_sampler.h
@@ -2,38 +2,39 @@
 #ifndef DOLEN_COMMON_SAMPLER_H
 #define DOLEN_COMMON_SAMPLER_H
 
+#include <stdint.h>
+
 
 typedef struct {
     float prob;
-    int index;
+    int32_t index;
 } prob_index;
 
 typedef struct {
-    int vocab_size;
+    int32_t vocab_size;
     prob_index *_prob_index;
     float temp;
-    int top_k;
+    int32_t top_k;
     float top_p;
-    unsigned long long rng_state;
+    uint64_t rng_state;
 } sampler;
 
 
-int sample_argmax(float *_probs, int n);
+int32_t sample_argmax(float *_probs, int32_t n);
 
-unsigned int random_u32(unsigned long long *_state);
+uint32_t random_u32(uint64_t *_state);
 
-float random_f32(unsigned long long *_state);
+float random_f32(uint64_t *_state);
 
-int sample_mult(float *_probs, int n, float coin);
+int32_t sample_mult(float *_probs, int32_t n, float coin);
 
-int compare_prob(const void *_a, const void *_b);
+int32_t compare_prob(const void *_a, const void *_b);
 
-int sample_top(float *_probs, int n, int top_k, float top_p, prob_index *_prob_index, float coin);
+int32_t sample_top(float *_probs, int32_t n, int32_t top_k, float top_p, prob_index *_prob_index, float coin);
 
-int sample(sampler *_sampler, float *_logits);
+int32_t sample(sampler *_sampler, float *_logits);
 
-void build_sampler(sampler *_sampler, int vocab_size, float temp, int top_k, float top_p,
-        unsigned long long rng_seed);
+void build_sampler(sampler *_sampler, int32_t vocab_size, float temp, int32_t top_k, float top_p, uint64_t rng_seed);
 
 void free_sampler(sampler *_sampler);
