+++ dolen_ig4_1.c
@@ -10,7 +10,7 @@ static const chat_template CHAT_TEMPLATE_IG4_1 = {
 };
 
 
-int load_quantized_ig4_1(const char *_file_path_s, IG4_1 *_model, int seq_n_max) {
+int32_t load_quantized_ig4_1(const char *_file_path_s, IG4_1 *_model, int32_t seq_n_max) {
     FILE *_file = fopen(_file_path_s, "rb");
     if (! _file) {
         log_msg(stderr, "ERROR: Failed to open %s for reading\n", _file_path_s);
@@ -74,7 +74,7 @@ int load_quantized_ig4_1(const char *_file_path_s, IG4_1 *_model, int seq_n_max)
         fclose(_file);
         return -1;
     }
-    for (int i = 0; i < _config->n_layer; i++) {
+    for (int32_t i = 0; i < _config->n_layer; i++) {
         read_qt(_file, &(_weights->_rms_att_weight[i]));
     }
 
@@ -83,7 +83,7 @@ int load_quantized_ig4_1(const char *_file_path_s, IG4_1 *_model, int seq_n_max)
     _weights->_wv = (qtensor *)a_calloc((size_t)_config->n_layer * sizeof(qtensor));
     _weights->_wo = (qtensor *)a_calloc((size_t)_config->n_layer * sizeof(qtensor));
 
-    for (int i = 0; i < _config->n_layer; i++) {
+    for (int32_t i = 0; i < _config->n_layer; i++) {
         read_qt(_file, &(_weights->_wq[i]));
         read_qt(_file, &(_weights->_wk[i]));
         read_qt(_file, &(_weights->_wv[i]));
@@ -96,14 +96,14 @@ int load_quantized_ig4_1(const char *_file_path_s, IG4_1 *_model, int seq_n_max)
         fclose(_file);
         return -1;
     }
-    for (int i = 0; i < _config->n_layer; i++) {
+    for (int32_t i = 0; i < _config->n_layer; i++) {
         read_qt(_file, &(_weights->_rms_ffn_weight[i]));
     }
 
     _weights->_w1 = (qtensor *)a_calloc((size_t)_config->n_layer * sizeof(qtensor));
     _weights->_w2 = (qtensor *)a_calloc((size_t)_config->n_layer * sizeof(qtensor));
     _weights->_w3 = (qtensor *)a_calloc((size_t)_config->n_layer * sizeof(qtensor));
-    for (int i = 0; i < _config->n_layer; i++) {
+    for (int32_t i = 0; i < _config->n_layer; i++) {
         read_qt(_file, &(_weights->_w1[i]));
         read_qt(_file, &(_weights->_w2[i]));
         read_qt(_file, &(_weights->_w3[i]));
@@ -124,17 +124,17 @@ int load_quantized_ig4_1(const char *_file_path_s, IG4_1 *_model, int seq_n_max)
     return 0;
 }
 
-void forward_ig4_1_attention_layer(IG4_1 *_model, int l, int pos) {
+void forward_ig4_1_attention_layer(IG4_1 *_model, int32_t l, int32_t pos) {
     config_ig4_1 *_config = &(_model->config);
     weights_ig4_1 *_weights = &(_model->weights);
     state_ig4_1 *_state = &(_model->state);
     float *_x = _state->_x;
-    int dim = _config->dim;
-    int head_size = _config->d_head > 0 ? _config->d_head : dim / _config->n_heads;
-    int kv_dim = _config->n_kv_heads * head_size;
-    int attn_out_dim = _config->n_heads * head_size;
-    int kv_mul = _config->n_heads / _config->n_kv_heads;
-    long long loff = (long long)l * _config->seq_len * kv_dim;
+    int32_t dim = _config->dim;
+    int32_t head_size = _config->d_head > 0 ? _config->d_head : dim / _config->n_heads;
+    int32_t kv_dim = _config->n_kv_heads * head_size;
+    int32_t attn_out_dim = _config->n_heads * head_size;
+    int32_t kv_mul = _config->n_heads / _config->n_kv_heads;
+    int64_t loff = (int64_t)l * _config->seq_len * kv_dim;
     float eps = _config->rms_norm_eps;
     float *_key_cache_row = _state->_key_cache + loff + pos * kv_dim;
     float *_value_cache_row = _state->_value_cache + loff + pos * kv_dim;
@@ -147,16 +147,16 @@ void forward_ig4_1_attention_layer(IG4_1 *_model, int l, int pos) {
     matmul_qq(_state->_k, &(_state->xq), &(_weights->_wk[l]));
     matmul_qq(_state->_v, &(_state->xq), &(_weights->_wv[l]));
 
-    int rotary_dim = head_size;
+    int32_t rotary_dim = head_size;
 
     if (_state->_cos_cache) {
         float *_cos_row = _state->_cos_cache + pos * rotary_dim;
         float *_sin_row = _state->_sin_cache + pos * rotary_dim;
 
 #pragma omp parallel for
-        for (int h = 0; h < _config->n_heads; h++) {
+        for (int32_t h = 0; h < _config->n_heads; h++) {
             float *_q = _state->_q + h * head_size;
-            for (int i = 0; i < rotary_dim / 2; i++) {
+            for (int32_t i = 0; i < rotary_dim / 2; i++) {
                 float c = _cos_row[i], sn = _sin_row[i];
                 float q0 = _q[i], q1 = _q[i + rotary_dim / 2];
                 _q[i] = q0 * c - q1 * sn;
@@ -165,9 +165,9 @@ void forward_ig4_1_attention_layer(IG4_1 *_model, int l, int pos) {
         }
 
 #pragma omp parallel for
-        for (int h = 0; h < _config->n_kv_heads; h++) {
+        for (int32_t h = 0; h < _config->n_kv_heads; h++) {
             float *_k = _state->_k + h * head_size;
-            for (int i = 0; i < rotary_dim / 2; i++) {
+            for (int32_t i = 0; i < rotary_dim / 2; i++) {
                 float c = _cos_row[i], sn = _sin_row[i];
                 float k0 = _k[i], k1 = _k[i + rotary_dim / 2];
                 _k[i] = k0 * c - k1 * sn;
@@ -185,16 +185,16 @@ void forward_ig4_1_attention_layer(IG4_1 *_model, int l, int pos) {
     }
 
 #pragma omp parallel for
-    for (int h = 0; h < _config->n_heads; h++) {
+    for (int32_t h = 0; h < _config->n_heads; h++) {
         float *_q = _state->_q + h * head_size;
         float *_att = _state->_att + h * _config->seq_len;
 
-        for (int t = 0; t <= pos; t++) {
+        for (int32_t t = 0; t <= pos; t++) {
             float *_k = _state->_key_cache + loff + t * kv_dim + (h / kv_mul) * head_size;
             float score = 0.0f;
 
 #pragma omp simd reduction(+ : score)
-            for (int i = 0; i < head_size; i++) {
+            for (int32_t i = 0; i < head_size; i++) {
                 score += _q[i] * _k[i];
             }
             _att[t] = score * attn_scale;
@@ -205,12 +205,12 @@ void forward_ig4_1_attention_layer(IG4_1 *_model, int l, int pos) {
         float *_xb = _state->_xb + h * head_size;
         memset(_xb, 0, head_size * sizeof(float));
 
-        for (int t = 0; t <= pos; t++) {
+        for (int32_t t = 0; t <= pos; t++) {
             float *_v = _state->_value_cache + loff + t * kv_dim + (h / kv_mul) * head_size;
             float a = _att[t];
 
 #pragma omp simd
-            for (int i = 0; i < head_size; i++) {
+            for (int32_t i = 0; i < head_size; i++) {
                 _xb[i] += a * _v[i];
             }
         }
@@ -220,18 +220,18 @@ void forward_ig4_1_attention_layer(IG4_1 *_model, int l, int pos) {
     matmul_qq(_state->_xb2, &(_state->xq), &(_weights->_wo[l]));
 
     float res_mult = _config->residual_multiplier;
-    for (int i = 0; i < dim; i++) {
+    for (int32_t i = 0; i < dim; i++) {
         _x[i] += _state->_xb2[i] * res_mult;
     }
 }
 
-void forward_ig4_1_mlp_layer(IG4_1 *_model, int l) {
+void forward_ig4_1_mlp_layer(IG4_1 *_model, int32_t l) {
     config_ig4_1 *_config = &(_model->config);
     weights_ig4_1 *_weights = &(_model->weights);
     state_ig4_1 *_state = &(_model->state);
     float *_x = _state->_x;
-    int dim = _config->dim;
-    int hidden_dim = _config->n_mlp;
+    int32_t dim = _config->dim;
+    int32_t hidden_dim = _config->n_mlp;
     float eps = _config->rms_norm_eps;
     float *_rms_ffn_weight = (float *)_weights->_rms_ffn_weight[l]._data;
 
@@ -242,7 +242,7 @@ void forward_ig4_1_mlp_layer(IG4_1 *_model, int l) {
     matmul_qq(_state->_hb2, &(_state->xq), &(_weights->_w3[l]));
 
 #pragma omp parallel for
-    for (int i = 0; i < hidden_dim; i++) {
+    for (int32_t i = 0; i < hidden_dim; i++) {
         float val = _state->_hb[i];
         val *= (1.0f / (1.0f + expf(-val)));
         val *= _state->_hb2[i];
@@ -253,29 +253,29 @@ void forward_ig4_1_mlp_layer(IG4_1 *_model, int l) {
     matmul_qq(_state->_xb, &(_state->hq), &(_weights->_w2[l]));
 
     float res_mult = _config->residual_multiplier;
-    for (int i = 0; i < dim; i++) {
+    for (int32_t i = 0; i < dim; i++) {
         _x[i] += _state->_xb[i] * res_mult;
     }
 }
 
-float *forward_ig4_1(IG4_1 *_model, int token, int pos) {
+float *forward_ig4_1(IG4_1 *_model, int32_t token, int32_t pos) {
     config_ig4_1 *_config = &(_model->config);
     weights_ig4_1 *_weights = &(_model->weights);
     state_ig4_1 *_state = &(_model->state);
     float *_x = _state->_x;
-    int dim = _config->dim;
+    int32_t dim = _config->dim;
 
     dequantize_row(_x, &(_weights->embed_tokens_weight), token);
 
     float emb_mult = _config->embedding_multiplier;
     if (emb_mult != 1.0f) {
 #pragma omp simd
-        for (int i = 0; i < dim; i++) {
+        for (int32_t i = 0; i < dim; i++) {
             _x[i] *= emb_mult;
         }
     }
 
-    for (int l = 0; l < _config->n_layer; l++) {
+    for (int32_t l = 0; l < _config->n_layer; l++) {
         forward_ig4_1_attention_layer(_model, l, pos);
         forward_ig4_1_mlp_layer(_model, l);
     }
@@ -292,7 +292,7 @@ float *forward_ig4_1(IG4_1 *_model, int token, int pos) {
     float logit_scale = _config->logits_scaling;
     if ((logit_scale != 0.0f) && (logit_scale != 1.0f)) {
 #pragma omp simd
-        for (int i = 0; i < _config->vocab_size; i++) {
+        for (int32_t i = 0; i < _config->vocab_size; i++) {
             _state->_logits[i] /= logit_scale;
         }
     }
@@ -300,7 +300,7 @@ float *forward_ig4_1(IG4_1 *_model, int token, int pos) {
     return _state->_logits;
 }
 
-static float *forward_ig4_1_wrap(void *_model, int token, int pos) {
+static float *forward_ig4_1_wrap(void *_model, int32_t token, int32_t pos) {
     return forward_ig4_1((IG4_1 *)_model, token, pos);
 }
 
@@ -309,7 +309,7 @@ static void free_ig4_1_wrap(void *_model) {
     free(_model);
 }
 
-model_iface *init_ig4_1(const char *_model_path_s, int seq_n_max, bool think_) {
+model_iface *init_ig4_1(const char *_model_path_s, int32_t seq_n_max, bool think_) {
     IG4_1 *_model = a_calloc(1 * sizeof(IG4_1));
 
     if (think_) {
